[REFACTOR] Move clipboard copy to an utility method
